Passes Unit Tests but Issue Remains
The software is working. Had to address the fact that some of the lists
in the report had a semi-arbitrary order so I ordered them. Also,
removed example files of .tar.zip as these were not in the example
output (may reconsider since these can be created but expected to be
common)